refactor: refactor comment directory for CQRS
