+++ backend/src/main/java/elbin_bank/issue_tracker/comment/application/query/dto/CommentDto.java
@@ -1,12 +1,14 @@
 package elbin_bank.issue_tracker.comment.application.query.dto;
 
+import elbin_bank.issue_tracker.user.infrastructure.query.projection.UserProjection;
+
 import java.time.LocalDateTime;
 
-public record CommentsSummaryDto(
+public record CommentDto(
         Long id,
-        UserDto author,
+        UserProjection author,
         String content,
         LocalDateTime createdAt,
         LocalDateTime updatedAt
-        ) {
+) {
 }
