+++ backend/src/main/java/elbin_bank/issue_tracker/comment/presentation/query/CommentQueryController.java
@@ -1,7 +1,8 @@
-package elbin_bank.issue_tracker.comment.presentation;
+package elbin_bank.issue_tracker.comment.presentation.query;
 
 import elbin_bank.issue_tracker.comment.application.query.CommentQueryService;
 import elbin_bank.issue_tracker.comment.application.query.dto.CommentsResponseDto;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,19 +11,16 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/v1/issues")
+@RequiredArgsConstructor
 public class CommentQueryController {
 
     private final CommentQueryService commentQueryService;
 
-    public CommentQueryController(CommentQueryService commentQueryService) {
-        this.commentQueryService = commentQueryService;
-    }
-
     @GetMapping("/{id}/comments")
     public ResponseEntity<CommentsResponseDto> getComments(@PathVariable long id) {
-        try{
+        try {
             return ResponseEntity.ok(commentQueryService.findByIssueId(id));
-        }catch (IllegalArgumentException e){
+        } catch (IllegalArgumentException e) {
             return ResponseEntity.notFound().build();
         }
     }
